+++ src/plugins/dev-mode/check-schema.ts
@@ -168,7 +168,12 @@ export function validateFieldsDeep(rxJsonSchema: RxJsonSchema<any>): true {
         if (!currentObj || typeof currentObj !== 'object') return;
         Object.keys(currentObj).forEach(attributeName => {
             const schemaObj = currentObj[attributeName];
-            if (!currentObj.properties && schemaObj && typeof schemaObj === 'object') {
+            if (
+                !currentObj.properties &&
+                schemaObj &&
+                typeof schemaObj === 'object' &&
+                !currentPath.endsWith('oneOf')
+            ) {
                 checkField(
                     attributeName,
                     schemaObj,